refactor(tools): move tools into script for general purpose
