+++ .eslintrc
@@ -6,7 +6,8 @@
         ],
         "quotes": [
             2,
-            "single"
+            "single",
+            { "allowTemplateLiterals": true }
         ],
         "linebreak-style": [
             2,
