+++ .make-packages.js
@@ -5,7 +5,7 @@ let fs = require('fs-extra');
 let mkdirp = require('mkdirp');
 let path = require('path');
 let klawSync = require('klaw-sync');
-let licenseTool = require('./tools/add-license-to-file');
+let licenseTool = require('./scripts/add-license-to-file');
 let addLicenseToFile = licenseTool.addLicenseToFile;
 let addLicenseTextToFile = licenseTool.addLicenseTextToFile;
 let bo = null;
