+++ esdoc.json
@@ -18,7 +18,7 @@
   ],
   "index": "./doc/index.md",
   "plugins": [
-    {"name": "./tools/custom-esdoc-plugin.js"}
+    {"name": "./scripts/custom-esdoc-plugin.js"}
   ],
   "manual": {
     "asset": "./doc/asset",
