+++ package.json
@@ -82,9 +82,9 @@
     "build_esm5_for_rollup": "npm-run-all clean_dist_esm5_for_rollup compile_dist_esm5_for_rollup",
     "build_esm2015": "npm-run-all clean_dist_esm2015 compile_dist_esm2015",
     "build_esm2015_for_docs": "npm-run-all clean_dist_esm2015 compile_dist_esm2015_for_docs",
-    "build_closure_core": "node ./tools/make-closure-core.js",
-    "build_global": "npm-run-all clean_dist_global build_esm5_for_rollup && mkdirp ./dist/global && node ./tools/make-umd-bundle.js && npm-run-all build_closure_core clean_dist_esm5_for_rollup",
-    "build_umd": "npm-run-all clean_dist_global && mkdirp ./dist/global && node ./tools/make-umd-bundle.js && npm-run-all build_closure_core",
+    "build_closure_core": "node ./scripts/make-closure-core.js",
+    "build_global": "npm-run-all clean_dist_global build_esm5_for_rollup && mkdirp ./dist/global && node ./scripts/make-umd-bundle.js && npm-run-all build_closure_core clean_dist_esm5_for_rollup",
+    "build_umd": "npm-run-all clean_dist_global && mkdirp ./dist/global && node ./scripts/make-umd-bundle.js && npm-run-all build_closure_core",
     "build_perf": "webdriver-manager update && npm-run-all build_cjs build_global perf",
     "build_docs": "npm-run-all build_global build_esm2015_for_docs build_cjs tests2png decision_tree_widget && esdoc -c esdoc.json && npm-run-all clean_dist_esm2015",
     "build_spec_browser": "webpack --config spec/support/webpack.mocha.config.js",
