+++ scripts/add-license-to-file.js
@@ -1,12 +1,16 @@
+/*eslint-env es6*/
+
 var fs = require('fs');
 var dist = require('minimist')(process.argv).dist;
 var license = require('minimist')(process.argv).license;
 
-function addLicenseToFile (license, destination) {
+// eslint-disable-next-line
+function addLicenseToFile(license, destination) {
   if (!license) {
     throw new Error('license path is required as 1st argument');
   }
 
+  // eslint-disable-next-line
   addLicenseTextToFile(fs.readFileSync(license).toString(), destination);
 }
